+++ mod/Scripting/Vixy_PlayerParts.cs
@@ -67,6 +67,7 @@ namespace QudExpandedCE
                 player.RequirePart<Vixy_TradeOffer>();
                 player.RequirePart<Vixy_LiquidGather>();
                 player.RequirePart<Vixy_MerchantOwnership>();
+                player.RequirePart<Vixy_WeaponWear>();
             }
         }
     }